Extract GitignoreRules from GitignoreManager for a testable public API
