+++ spec/spec_helper.rb
@@ -23,6 +23,7 @@ require_relative '../lib/ghb/dockerhub_manager'
 require_relative '../lib/ghb/file_scanner'
 require_relative '../lib/ghb/github_api_client'
 require_relative '../lib/ghb/gitignore_manager'
+require_relative '../lib/ghb/gitignore_rules'
 require_relative '../lib/ghb/language_job_builder'
 require_relative '../lib/ghb/licenses_job_builder'
 require_relative '../lib/ghb/linter_job_builder'
